Изменения с версии 0.5.4:

Абитуриенты и списки:
- DK-49: над таблицей списка абитуриентов добавлены кнопки-переключатели
  представлений — «Общая информация» и «Приоритетные специальности» (ФИО + теги
  специальностей по приоритетам в столбцах 1, 2, 3…). Активное — primary,
  остальные — secondary. Данные неактивного представления не держатся в памяти;
  каждое грузится своим методом (для второго — облегчённая выборка). У каждого
  представления свои независимые фильтры; фильтр по приоритетной специальности
  означает специальность с приоритетом №1 (дублирующий фильтр по любому приоритету
  убран).
- DK-45: поле «Код ЄДБО» в карточке абитуриента переименовано в «ID особи в ЄДБО»
  (имя колонки в БД без изменений); в список абитуриентов добавлен столбец «Група»
  с прочерком для неназначенной группы и серверной сортировкой по названию
  (UA_CI-collation).

Приёмная кампания и группы:
- DK-48: переработка автоформирования и работы с экзаменационными группами. Имя
  группы: тег берётся из базы поступления, а если он пустой — из формы обучения.
  Тип существующей группы определяется только по названию; абитуриенты без
  специальностей в группы не попадают. Новая глобальная настройка «Максимум
  абитуриентов в группе» применяется везде (автоформирование, автоподбор,
  предупреждение о лимите при ручном выборе). Автоподбор группы в карточке
  абитуриента и автоподбор абитуриентов под конкретную группу на странице её
  редактирования. Сортировка списка групп по названию и количеству абитуриентов;
  код автоформирования унифицирован для bulk- и single-режимов.

Рейтинг:
- DK-47: флаг is_counted_in_rating у предметов ЗНО (по умолчанию выкл) — в сумму
  баллов рейтинга и в столбцы DOCX входят только отмеченные предметы (напр. без
  отдельных компонентов НМТ). Оценки стали дробными (results_zno.points/points_raw
  и rating_entries.total_points → float, вывод через format_grade). Калькулятор
  комплексного балла (среднее/сумма компонентов) в диалоге результата ЗНО.
  Миграция d4e5f6a7b8c9.

Отчётность и документы:
- DK-30: новая страница журнала регистрации заявлений (/reporting/registration_journal)
  — живой просмотр формы МОН с фильтром по кампании и дате приёма + выгрузка DOCX
  по шаблону. Права report_journal:view / report_journal:docx.
- DK-46: экспорт графика экзаменов в DOCX по шаблону вместо страницы печати
  (удалены маршрут ENTRANT_EXAM_PRINT и print-view). Диалог выбора экзаменов;
  рендер и загрузка в фоновом потоке.

Прочее и инфраструктура:
- DK-50: все метки времени (создание/редактирование карточек, формирование отчётов,
  токены сессии) переведены на киевское время (EET/EEST) через
  utils/clock.py:now_local() вместо datetime.now(), отдававшего UTC в контейнере.
  Добавлена зависимость tzdata; применяется к новым записям.
- Версия проекта в pyproject.toml и uv.lock приведена к 0.6.
- Пересборка скомпилированного фронтенда для nginx.

+++ Dekanat/services/result_zno.py
@@ -23,8 +23,8 @@ class ResultZnoService:
         self,
         id_items_zno: int,
         id_person: int,
-        points: Optional[int],
-        points_raw: Optional[int] = None,
+        points: Optional[float],
+        points_raw: Optional[float] = None,
     ) -> None:
         """Якщо `points is None` — видаляє існуючу оцінку. Інакше — створює/оновлює.
         `points` — підсумковий (домножений) бал, `points_raw` — сирий (введений)."""
